style: make bot happy

+++ interfaces/curve/VoteEscrow.sol
@@ -3,7 +3,9 @@
 pragma solidity ^0.5.17;
 
 interface VoteEscrow {
-    function create_lock(uint, uint) external;
-    function increase_amount(uint) external;
+    function create_lock(uint256, uint256) external;
+
+    function increase_amount(uint256) external;
+
     function withdraw() external;
 }
