+++ interfaces/yearn/Proxy.sol
@@ -3,6 +3,11 @@
 pragma solidity ^0.5.17;
 
 interface Proxy {
-    function execute(address to, uint value, bytes calldata data) external returns (bool, bytes memory);
-    function increaseAmount(uint) external;
+    function execute(
+        address to,
+        uint256 value,
+        bytes calldata data
+    ) external returns (bool, bytes memory);
+
+    function increaseAmount(uint256) external;
 }
