+++ interfaces/yearn/VoterProxy.sol
@@ -3,9 +3,17 @@
 pragma solidity ^0.5.17;
 
 interface VoterProxy {
-    function withdraw(address _gauge, address _token, uint _amount) external returns (uint);
-    function balanceOf(address _gauge) external view returns (uint);
-    function withdrawAll(address _gauge, address _token) external returns (uint);
+    function withdraw(
+        address _gauge,
+        address _token,
+        uint256 _amount
+    ) external returns (uint256);
+
+    function balanceOf(address _gauge) external view returns (uint256);
+
+    function withdrawAll(address _gauge, address _token) external returns (uint256);
+
     function deposit(address _gauge, address _token) external;
+
     function harvest(address _gauge) external;
 }
